+++ documentdb_tests/compatibility/tests/core/operator/accumulators/min/test_accumulator_min_group_input_and_return.py
@@ -1,0 +1,180 @@
+"""Tests for $min accumulator — expression arguments and return type verification ($group)."""
+
+from __future__ import annotations
+
+from typing import Any
+
+import pytest
+from bson import Decimal128, Int64
+
+from documentdb_tests.compatibility.tests.core.operator.accumulators.utils import (
+    AccumulatorTestCase,
+)
+from documentdb_tests.framework.assertions import assertSuccess
+from documentdb_tests.framework.executor import execute_command
+from documentdb_tests.framework.parametrize import pytest_params
+from documentdb_tests.framework.test_constants import DATE_EPOCH, DATE_Y2K
+
+
+def _group_min(accumulator: Any) -> list[dict[str, Any]]:
+    """Build a $group pipeline for $min."""
+    return [{"$group": {"_id": None, "result": {"$min": accumulator}}}]
+
+
+def _run_accumulator(collection, test_case: AccumulatorTestCase):
+    """Insert docs and run the pipeline."""
+    if test_case.docs:
+        collection.insert_many(test_case.docs)
+    return execute_command(
+        collection,
+        {"aggregate": collection.name, "pipeline": test_case.pipeline, "cursor": {}},
+    )
+
+
+# ---------------------------------------------------------------------------
+# Property [Expression Argument Tests]: $min accumulator accepts various
+# expression types as its operand.
+# ---------------------------------------------------------------------------
+MIN_EXPRESSION_ARGUMENT_TESTS: list[AccumulatorTestCase] = [
+    AccumulatorTestCase(
+        "input_field_path",
+        docs=[{"v": 10}, {"v": 5}, {"v": 20}],
+        pipeline=_group_min("$v"),
+        expected=[{"_id": None, "result": 5}],
+        msg="$min should accept basic field reference",
+    ),
+    AccumulatorTestCase(
+        "input_nested_field",
+        docs=[{"a": {"b": 10}}, {"a": {"b": 5}}, {"a": {"b": 20}}],
+        pipeline=_group_min("$a.b"),
+        expected=[{"_id": None, "result": 5}],
+        msg="$min should accept nested document field path",
+    ),
+    AccumulatorTestCase(
+        "input_literal",
+        docs=[{"v": 1}, {"v": 2}, {"v": 3}],
+        pipeline=_group_min(42),
+        expected=[{"_id": None, "result": 42}],
+        msg="$min should accept constant literal (same for all docs)",
+    ),
+    AccumulatorTestCase(
+        "input_expression",
+        docs=[{"price": 10, "qty": 2}, {"price": 5, "qty": 3}, {"price": 20, "qty": 1}],
+        pipeline=_group_min({"$multiply": ["$price", "$qty"]}),
+        expected=[{"_id": None, "result": 15}],
+        msg="$min should accept computed expression as operand",
+    ),
+    AccumulatorTestCase(
+        "input_cond_remove",
+        docs=[{"v": -1}, {"v": 5}, {"v": -2}, {"v": 10}],
+        pipeline=_group_min({"$cond": [{"$gt": ["$v", 0]}, "$v", "$$REMOVE"]}),
+        expected=[{"_id": None, "result": 5}],
+        msg="$min should accept conditional expression with $$REMOVE",
+    ),
+    AccumulatorTestCase(
+        "input_null_literal",
+        docs=[{"v": 1}, {"v": 2}],
+        pipeline=_group_min(None),
+        expected=[{"_id": None, "result": None}],
+        msg="$min should return null when accumulator is null literal",
+    ),
+]
+
+# ---------------------------------------------------------------------------
+# Property [Return Type Verification]: $min preserves the BSON type of the minimum value.
+# ---------------------------------------------------------------------------
+MIN_RETURN_TYPE_TESTS: list[AccumulatorTestCase] = [
+    AccumulatorTestCase(
+        "return_type_int32",
+        docs=[{"v": 10}, {"v": 20}, {"v": 30}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": 10, "type": "int"}],
+        msg="$min should preserve int32 return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_int64",
+        docs=[{"v": Int64(10)}, {"v": Int64(20)}, {"v": Int64(30)}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": Int64(10), "type": "long"}],
+        msg="$min should preserve int64 return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_double",
+        docs=[{"v": 1.5}, {"v": 2.5}, {"v": 3.5}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": 1.5, "type": "double"}],
+        msg="$min should preserve double return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_decimal",
+        docs=[{"v": Decimal128("1.5")}, {"v": Decimal128("2.5")}, {"v": Decimal128("3.5")}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": Decimal128("1.5"), "type": "decimal"}],
+        msg="$min should preserve Decimal128 return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_string",
+        docs=[{"v": "apple"}, {"v": "banana"}, {"v": "cherry"}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": "apple", "type": "string"}],
+        msg="$min should preserve string return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_boolean",
+        docs=[{"v": True}, {"v": False}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": False, "type": "bool"}],
+        msg="$min should preserve boolean return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_date",
+        docs=[{"v": DATE_EPOCH}, {"v": DATE_Y2K}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": DATE_EPOCH, "type": "date"}],
+        msg="$min should preserve date return type",
+    ),
+    AccumulatorTestCase(
+        "return_type_null_all",
+        docs=[{"v": None}, {"v": None}],
+        pipeline=[
+            {"$group": {"_id": None, "result": {"$min": "$v"}}},
+            {"$project": {"_id": 0, "value": "$result", "type": {"$type": "$result"}}},
+        ],
+        expected=[{"value": None, "type": "null"}],
+        msg="$min should return null type when all values are null",
+    ),
+]
+
+
+# ---------------------------------------------------------------------------
+# Combined success tests
+# ---------------------------------------------------------------------------
+MIN_GROUP_INPUT_AND_RETURN_SUCCESS_TESTS = MIN_EXPRESSION_ARGUMENT_TESTS + MIN_RETURN_TYPE_TESTS
+
+
+@pytest.mark.parametrize("test_case", pytest_params(MIN_GROUP_INPUT_AND_RETURN_SUCCESS_TESTS))
+def test_accumulator_min_group_input_and_return(collection, test_case: AccumulatorTestCase):
+    """Test $min accumulator expression arguments and return type verification with $group."""
+    result = _run_accumulator(collection, test_case)
+    assertSuccess(result, test_case.expected, msg=test_case.msg)
